Refactor: move files to new dir and update imports
