--- frontend/src/Components/NavLink/__test__/NavLink.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import NavLink from '../NavLink';
-
-describe('NavLink component', () => {
-  it('Does something...', () => {
-
-  });
-});
