--- frontend/src/Components/Navbar/__test__/Navbar.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import Navbar from '../Navbar';
-
-describe('Navbar component', () => {
-  it('Does something...', () => {
-
-  });
-});
