--- frontend/src/Components/SearchBar/__test__/SearchBar.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import SearchBar from '../SearchBar';
-
-describe('SearchBar component', () => {
-  it('Does something...', () => {
-
-  });
-});
